requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Django==5.1.4
elevate==0.1.3
ffmpeg-python==0.2.0
filelock==3.16.1
fonttools==4.55.2
fsspec==2024.10.0
future==1.0.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
lapx==0.5.11.post1
MarkupSafe==3.0.2
matplotlib==3.9.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
py-cpuinfo==9.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.6
six==1.17.0
sqlparse==0.5.2
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.40
ultralytics-thop==2.0.12
urllib3==2.2.3
